+++ packages/media-parser/src/create/iso-base-media/create-stbl.ts
@@ -1,14 +1,14 @@
 import {combineUint8Arrays} from '../../boxes/webm/make-header';
 import type {SamplePosition} from '../../get-sample-positions';
+import {createCttsBox} from './mdia/stbl/create-ctts';
+import {createStcoAtom} from './mdia/stbl/create-stco';
+import {createStsc} from './mdia/stbl/create-stsc';
+import {createStss} from './mdia/stbl/create-stss';
+import {createStsz} from './mdia/stbl/create-stsz';
+import {createSttsAtom} from './mdia/stbl/create-stts';
+import type {Avc1Data} from './mdia/stbl/stsd/create-avc1';
+import {createStsdData} from './mdia/stbl/stsd/create-avc1';
 import {addSize, stringsToUint8Array} from './primitives';
-import {createCttsBox} from './stbl/create-ctts';
-import {createStcoAtom} from './stbl/create-stco';
-import {createStsc} from './stbl/create-stsc';
-import {createStss} from './stbl/create-stss';
-import {createStsz} from './stbl/create-stsz';
-import {createSttsAtom} from './stbl/create-stts';
-import type {Avc1Data} from './stbl/stsd/create-avc1';
-import {createStsdData} from './stbl/stsd/create-avc1';
 
 export const createStbl = ({
 	samplePositions,
